requirements.txt
handright==8.0.0
Pillow==9.3.0